Code.c
Code for HangMan game in C language.

This hangman game in c code is about guessing letters (A-Z) to form the words. A hangman game is a common word guessing game in which the player must guess one letter at a time to complete a missing word.

This hangman game source code in c language ends after a certain number of incorrect guesses, and the player loses. If the player correctly recognizes all of the letters in the missing title, the game is also over. In this project design is straightforward and clean, making it easy for users to learn, use, and navigate.